requirements.txt
langchain
langchain-community
langchain-core
langchain-text-splitters
langchain-huggingface
huggingface_hub
sentence-transformers
faiss-cpu
streamlit
ragas
pymupdf
python-dotenv
transformers
torch